refactor: RecentlyViewedArtworks section
